restructure the folder stucture
