+++ Project/src/main/java/model/parkModel/ParkModel.java
@@ -1,4 +1,4 @@
-package model;
+package model.parkModel;
 
 public class ParkModel {
