+++ Project/src/test/java/ParkModelTest.java
@@ -1,10 +1,7 @@
-
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
 import org.junit.Test;
 import static org.junit.Assert.*;
-import model.*;
+import model.parkModel.*;
+import model.vehicleModel.*;
 
 public class ParkModelTest {
